% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asylum_applications}
\alias{asylum_applications}
\title{Asylum applications by asylum-seekers}
\format{
A data frame with 85949 rows and 13 variables:
\describe{
  \item{\code{Year}}{character Year}
  \item{\code{CountryOriginCode}}{character Country of Origin Code isoA3}
  \item{\code{CountryAsylumCode}}{character Country of Asylum Code isoA3}
  \item{\code{CountryOriginName}}{character Country of Origin Name}
  \item{\code{CountryAsylumName}}{character Country of Asylum Name}
  \item{\code{ProcedureType}}{character Procedure Type}
  \item{\code{ProcedureName}}{character Procedure Name}
  \item{\code{ApplicationTypeCode}}{character Application Type Code}
  \item{\code{ApplicationType}}{character Application Type}
  \item{\code{ApplicationDataType}}{character Application Data Type}
  \item{\code{ApplicationData}}{character Application Data}
  \item{\code{ApplicationAveragePersonsPerCase}}{integer Application Average Persons Per Case}
  \item{\code{NumberApplications}}{integer Number of Applications} 
}
}
\source{
\url{https://data.humdata.org/dataset/unhcr-population-data-for-world}
}
\usage{
asylum_applications
}
\description{
Dataset collated by UNHCR, containing information about the number of applications for asylum residing in World. Data is available since 2000
}
\keyword{datasets}
