% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asylum_decisions}
\alias{asylum_decisions}
\title{Asylum decisions taken on asylum claims of asylum-seekers}
\format{
A data frame with 87495 rows and 16 variables:
\describe{
  \item{\code{Year}}{character Year}
  \item{\code{CountryOriginCode}}{character Country of Origin Code isoA3}
  \item{\code{CountryAsylumCode}}{character Country of Asylum Code isoA3}
  \item{\code{CountryOriginName}}{character Country of Origin Name}
  \item{\code{CountryAsylumName}}{character Country of Asylum Name}
  \item{\code{ProcedureType}}{character Procedure Type}
  \item{\code{ProcedureName}}{character Procedure Name}
  \item{\code{DecisionTypeCode}}{character Decision Type Code}
  \item{\code{DecisionTypeName}}{character Decision Type Name}
  \item{\code{DecisionDataType}}{character Decision Data Type}
  \item{\code{DecisionData}}{character Decision Data}
  \item{\code{DecisionsAveragePersonsPerCase}}{integer Decisions Average Persons Per Case}
  \item{\code{Recognized}}{integer Recognized}
  \item{\code{ComplementaryProtection}}{integer Complementary Protection}
  \item{\code{OtherwiseClosed}}{integer Otherwise Closed}
  \item{\code{Rejected}}{integer Rejected}
  \item{\code{Total Decided}}{integer Total Decided} 
}
}
\source{
\url{https://data.humdata.org/dataset/unhcr-population-data-for-world}
}
\usage{
asylum_decisions
}
\description{
Data collated by UNHCR about asylum decisions taken on asylum claims of asylum-seekers residing in World. Data is available since 2000.
}
\keyword{datasets}
