% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asylum_decisions_long}
\alias{asylum_decisions_long}
\title{Asylum decisions taken on asylum claims of asylum-seekers}
\format{
A data frame with 160,334 rows and 13 variables:
\describe{
  \item{\code{Year}}{character Year}
  \item{\code{CountryOriginCode}}{character Country of Origin Code isoA3}
  \item{\code{CountryAsylumCode}}{character Country of Asylum Code isoA3}
  \item{\code{CountryOriginName}}{character Country of Origin Name}
  \item{\code{CountryAsylumName}}{character Country of Asylum Name}
  \item{\code{ProcedureType}}{character Procedure Type}
  \item{\code{ProcedureName}}{character Procedure Name}
  \item{\code{DecisionTypeCode}}{character Decision Type Code}
  \item{\code{DecisionTypeName}}{character Decision Type Name}
  \item{\code{DecisionData}}{character Decision Data}
  \item{\code{DecisionsAveragePersonsPerCase}}{integer Decisions Average Persons Per Case}
  \item{\code{Decision.output}}{integer COLUMN_DESCRIPTION}
  \item{\code{Value}}{double COLUMN_DESCRIPTION} 
}
}
\source{
\url{https://data.humdata.org/dataset/unhcr-population-data-for-world}
}
\usage{
asylum_decisions_long
}
\description{
Data collated by UNHCR about asylum decisions taken on asylum claims of asylum-seekers residing in World. Data is available since 2000.
}
\keyword{datasets}
