% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demographics}
\alias{demographics}
\title{Demographics and locations of forcibly displaced displaced and stateless persons}
\format{
A data frame with 182359 rows and 24 variables:
\describe{
  \item{\code{Year}}{character Year}
  \item{\code{CountryOriginCode}}{character Country of Origin Code isoA3}
  \item{\code{CountryAsylumCode}}{character Country of Asylum Code isoA3}
  \item{\code{CountryOriginName}}{character Country of Origin Name}
  \item{\code{CountryAsylumName}}{character Country of Asylum Name}
  \item{\code{Population Type}}{character Population Type: Population types are expressed in the data as follows:#'   
  *    `REF` – Refugee  
  *    `ROC` – People in refugee-like situation  
  *    `ASY` – Asylum seekers  
  *    `VDA` – Venezuelans displaced abroad  
  *    `IDP` – Internally displaced persons  
  *    `IOC` – People in IDP-like situation  
  *    `STA` – Stateless persons  
  *    `OOC` – Others of concern  }
  \item{\code{location}}{character location}
  \item{\code{urbanRural}}{character Urban  or Rural}
  \item{\code{accommodationType}}{character Accommodation Type}
  \item{\code{Female04}}{integer Female 0-4}
  \item{\code{Female511}}{integer Female 5-11}
  \item{\code{Female1217}}{integer Female 12-17}
  \item{\code{Female1859}}{integer Female 18-59}
  \item{\code{Female60ormore}}{integer Female 60 or more}
  \item{\code{FemaleUnknown}}{integer Female Unknown}
  \item{\code{FemaleTotal}}{integer Female Total}
  \item{\code{Male04}}{integer Male 0-4}
  \item{\code{Male511}}{integer Male 5-11}
  \item{\code{Male1217}}{integer Male 12-17}
  \item{\code{Male1859}}{integer Male 18-59}
  \item{\code{Male60ormore}}{integer Male 60 or more}
  \item{\code{MaleUnknown}}{integer Male Unknown}
  \item{\code{MaleTotal}}{integer Male}
  \item{\code{Total}}{integer Total} 
}
}
\source{
\url{https://data.humdata.org/dataset/unhcr-population-data-for-world}
}
\usage{
demographics
}
\description{
Data collated by UNHCR, containing information about forcibly displaced populations and stateless persons residing in World. It includes data on their demographics as well as sub-national locations. Data is available from 2001.
}
\keyword{datasets}
