% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{end_year_population_totals}
\alias{end_year_population_totals}
\title{end_year_population_totals}
\format{
A data frame with 103302 rows and 11 variables:
\describe{
  \item{\code{Year}}{character Year}
  \item{\code{CountryOriginCode}}{character Country of Origin Code isoA3}
  \item{\code{CountryAsylumCode}}{character Country of Asylum Code isoA3}
  \item{\code{CountryOriginName}}{character Country of Origin Name}
  \item{\code{CountryAsylumName}}{character Country of Asylum Name}
  \item{\code{Refugees}}{integer Number of Refugees}
  \item{\code{IDPs}}{integer Number of Internally Displaced Persons}
  \item{\code{AsylumSeekers}}{integer Number of Asylum-seekers}
  \item{\code{OthersOfConcern}}{integer Number of Others of Concern to UNHCR}
  \item{\code{StatelessPersons}}{integer Number of Stateless persons}
  \item{\code{VenezuelansDisplacedAbroad}}{integer Number of Venezuelans Displaced Abroad} 
}
}
\source{
\url{https://data.humdata.org/dataset/unhcr-population-data-for-world}
}
\usage{
end_year_population_totals
}
\description{
Data collated by UNHCR, containing end-year stock population figures for forcibly displaced and stateless persons residing in World. Data is available since 1951.
}
\keyword{datasets}
