% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{end_year_population_totals_long}
\alias{end_year_population_totals_long}
\title{End-year stock population figures for forcibly displaced displaced and stateless persons - Long Format}
\format{
A data frame with 153809 rows and 7 variables:
\describe{
  \item{\code{Year}}{character Year}
  \item{\code{CountryOriginCode}}{character Country of Origin Code isoA3}
  \item{\code{CountryAsylumCode}}{character Country of Asylum Code isoA3}
  \item{\code{CountryOriginName}}{character Country of Origin Name}
  \item{\code{CountryAsylumName}}{character Country of Asylum Name}
  \item{\code{Population.type}}{character Type of Solution }
  \item{\code{Population.type.label}}{character Type of Population }
  \item{\code{Population.type.label.short}}{character Type of Population - short label }
  \item{\code{Value}}{integer Number of person} 
}
}
\source{
\url{https://data.humdata.org/dataset/unhcr-population-data-for-world}
}
\usage{
end_year_population_totals_long
}
\description{
Data collated by UNHCR, containing end-year stock population figures for forcibly displaced and stateless persons residing in World. Data is available since 1951.
}
\keyword{datasets}
