% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imf}
\alias{imf}
\title{IMF GDP data}
\format{
A data frame with 193 rows and 8 variables:
\describe{
  \item{\code{ISO}}{character iso3 country code}
  \item{\code{Country}}{character Country name}
  \item{\code{Units}}{character Units of the data}
  \item{\code{Scale}}{logical}
  \item{\code{Country/Series-specific Notes}}{character Data source}
  \item{\code{2020}}{double GDP data for 2020}
  \item{\code{2021}}{character GDP data for 2021}
  \item{\code{Estimates Start After}}{double Estimate time} 
}
}
\source{
\url{https://www.imf.org}
}
\usage{
imf
}
\description{
GDP at constant price for 2020&2021 from IMF
}
\keyword{datasets}
