% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{locpcode}
\alias{locpcode}
\title{Geo-referential}
\format{
A data frame with 1653 rows and 7 variables:
\describe{
  \item{\code{asylum}}{character Country of Asylum - UNHCR 3 letters, aka UNHCRcode from reference}
  \item{\code{location}}{character Name o fthe locatin}
  \item{\code{latitude}}{character latitude - geographic coordinate}
  \item{\code{longitude}}{character longttude - geographic coordinate}
  \item{\code{location_pcode}}{character pcode used by https://gis.unhcr.org/arcgis/rest/services/population/ }
  \item{\code{location_level}}{double Admin level hierarchy}
  \item{\code{location_id}}{double location id} 
  \item{\code{iso_3}}{character Country of Asylum - Iso 3 letters} 
}
}
\usage{
locpcode
}
\description{
Location used at sub-national level to record population data
}
\details{
DETAILS
}
\keyword{datasets}
