% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctr_asylum.R
\name{plot_ctr_asylum}
\alias{plot_ctr_asylum}
\title{Asylum Applications & Decision over time}
\usage{
plot_ctr_asylum(
  year = 2022,
  country_asylum_iso3c = country_asylum_iso3,
  lag = 10
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{country_asylum_iso3c}{Character value with the ISO-3 character code of the Country of Asylum}

\item{lag}{Number of year to used as comparison base}
}
\value{
a ggplot2 object
}
\description{
This charts allow to visualize trends in terms of asylum applications, decision and refugee status recognition
}
\examples{
plot_ctr_asylum(year = 2022,
                country_asylum_iso3c = "ECU", 
                            lag = 10)
}
