% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctr_destination.R
\name{plot_ctr_destination}
\alias{plot_ctr_destination}
\title{Tree map of Population Groups within a country}
\usage{
plot_ctr_destination(
  year = 2021,
  country_origin_iso3c = country_origin_iso3c,
  pop_type = pop_type
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{country_origin_iso3c}{Character value with the ISO-3 character code of the Country of Origin}

\item{pop_type}{Vector of character values. Possible population type (e.g.: REF, IDP, ASY, OIP, OOC, STA)}
}
\value{
a ggplot2 object
}
\description{
Tree map of Population Groups within a country
}
\examples{
# 
plot_ctr_destination(year = 2021,
                     country_origin_iso3c = "COL",
                     pop_type = c("REF", "ASY")
         )

}
