% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctr_origin_history.R
\name{plot_ctr_origin_history}
\alias{plot_ctr_origin_history}
\title{Plot Origin History}
\usage{
plot_ctr_origin_history(
  year = 2024,
  lag = 5,
  country_asylum_iso3c,
  pop_type = c("REF", "ASY", "IDP", "OIP", "STA", "OOC"),
  otherprop = 0.02
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{lag}{Number of year to used as comparison base}

\item{country_asylum_iso3c}{Character value with the ISO-3 character code of the Country of Asylum}

\item{pop_type}{Vector of character values. Possible population type (e.g.: REF, IDP, ASY, OIP, OIP, OOC, STA)}

\item{otherprop}{value set by default to .02 - used to merge origin as "Other"}
}
\value{
a ggplot2 object
}
\description{
The functions gives a quick overview of the main source of displacement in terms of origin in the country
}
\examples{
plot_ctr_origin_history(
  year = 2024,
  lag = 5,
  country_asylum_iso3c = "MEX",
  pop_type = c(
    "REF",
    "ASY",
    "OIP",
    "IDP"
  ),
  otherprop = .02
)
}
