% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctr_population_type_abs.R
\name{plot_ctr_population_type_abs}
\alias{plot_ctr_population_type_abs}
\title{Main country of origin - Absolute value}
\usage{
plot_ctr_population_type_abs(
  year = 2021,
  country_asylum_iso3c = country_asylum_iso3c,
  top_n_countries = 9,
  pop_type = "REF",
  show_diff_label = TRUE
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{country_asylum_iso3c}{Character value with the ISO-3 character code of the Country of Asylum}

\item{top_n_countries}{Numeric value of number of main countries that the graph should display}

\item{pop_type}{Character value. Possible population type (e.g.: REF, IDP, ASY,   OIP, OOC, STA)}

\item{show_diff_label}{logical to indicate whether or not adding the the label displaying difference in percentage compared to the previous year}
}
\value{
a ggplot2 object
}
\description{
Main country of origin - Absolute value
}
\examples{
plot_ctr_population_type_abs(year = 2020,
                    country_asylum_iso3c = "USA",
                    top_n_countries = 4,
                    pop_type = "REF",
                    show_diff_label = FALSE
                    ) 

## Same with 9 top countries and Asylum-seekers included
plot_ctr_population_type_abs(year = 2020,
                    country_asylum_iso3c = "USA",
                    top_n_countries = 9,
                    pop_type = "ASY",
                    show_diff_label = TRUE
                    ) 

}
