% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctr_population_type_perc.R
\name{plot_ctr_population_type_perc}
\alias{plot_ctr_population_type_perc}
\title{Main country of origin - Percentage}
\usage{
plot_ctr_population_type_perc(
  year = 2021,
  country_asylum_iso3c = country_asylum_iso3c,
  top_n_countries = 9,
  pop_type = "REF"
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{country_asylum_iso3c}{Character value with the ISO-3 character code of the Country of Asylum}

\item{top_n_countries}{Numeric value of number of main countries that the graph should display}

\item{pop_type}{Character value. Possible population type (e.g.: REF, IDP, ASY, OIP, OOC, STA)}
}
\description{
Main country of origin - Percentage
}
\examples{

plot_ctr_population_type_perc(year = 2021,
                    country_asylum_iso3c = "BRA",
                    top_n_countries = 9,
                    pop_type = "REF"  ) 
}
