% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctr_process.R
\name{plot_ctr_process}
\alias{plot_ctr_process}
\title{Asylum Processing}
\usage{
plot_ctr_process(year = 2022, country_asylum_iso3c)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{country_asylum_iso3c}{Character value with the ISO-3 character code of the Country of Asylum}
}
\value{
a ggplot2 object
}
\description{
Displaying Asylum processing
}
\examples{
plot_ctr_process(year = 2022,   country_asylum_iso3c = "CHL")
}
