% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctr_recognition.R
\name{plot_ctr_recognition}
\alias{plot_ctr_recognition}
\title{Display A Chart with Refugee Recognition Rater}
\usage{
plot_ctr_recognition(
  year = 2022,
  country_asylum_iso3c,
  top_n_countries = 10,
  measure = "RefugeeRecognitionRate",
  order_by = "TotalDecided"
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{country_asylum_iso3c}{Character value with the ISO-3 character code of the Country of Asylum}

\item{top_n_countries}{Numeric value of number of main countries that the graph should display}

\item{measure}{this can be either:
* RefugeeRecognitionRate 
* TotalRecognitionRate}

\item{order_by}{this can be either:
* Recognized  
* ComplementaryProtection  
* TotalDecided}
}
\value{
a ggplot2 object
}
\description{
In the absence of an internationally agreed methodology for calculating recognition rates,
 UNHCR uses two rates to compute the proportion of refugee claims accepted during the year: 
 
  * The Refugee Recognition Rate divides the number of asylum-seekers granted Convention refugee
  status by the total number of accepted (Convention and, where relevant, complementary 
  protection) and rejected cases. 
  
  * The Total Recognition Rate divides the number of asylum-seekers granted Convention
   refugee status and complementary form of protection by the total number
   of accepted (Convention and, where relevant, complementary protection) and
    rejected cases. Non-substantive decisions are, to the extent possible, 
    excluded from both calculations. For the purpose of international comparability, 
    UNHCR only uses these two recognition rates and does not report nationally
     calculated rates.
     
     See https://www.unhcr.org/4ce531e09.pdf
}
\examples{
plot_ctr_recognition(year = 2022,
                     country_asylum_iso3c = "USA",
                     top_n_countries = 10, 
                     measure = "RefugeeRecognitionRate",
                     order_by = "TotalDecided" )
}
