% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctr_solution.R
\name{plot_ctr_solution}
\alias{plot_ctr_solution}
\title{Plotting solutions within a country}
\usage{
plot_ctr_solution(
  year = 2021,
  lag = 10,
  country_asylum_iso3c = country_asylum_iso3c,
  sol_type
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{lag}{Number of year to used as comparison base}

\item{country_asylum_iso3c}{Character value with the ISO-3 character code of the Country of Asylum}

\item{sol_type}{Vector of character values. Possible population type (e.g.:"NAT" "RST" "RET" "RDP")}
}
\value{
a ggplot2 object
}
\description{
As part of UNHCR standard approach, there are mostly 3 types of durable solutions for Refugees:
 - Voluntary Repatriation in origin country 
 - Naturalization in the host country
 - Resettlement in a third country
}
\examples{
plot_ctr_solution(year = 2021,
                  country_asylum_iso3c= "UGA",
                  lag = 10,
                  sol_type = c("NAT", "RST", "RET", "RDP"))
}
