% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_decrease.R
\name{plot_reg_decrease}
\alias{plot_reg_decrease}
\title{Plot Biggest decrease in Refugee Population}
\usage{
plot_reg_decrease(
  year = 2021,
  lag = 5,
  topn = 5,
  region = "The Americas",
  pop_type = c("REF", "ASY", "OIP")
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{lag}{Number of year to used as comparison base}

\item{topn}{how many top countries to show..}

\item{region}{Character value with the related UNHCR bureau - when left null,
it will display the whole world}

\item{pop_type}{Vector of character values. Possible population type
(e.g.: REF, IDP, ASY, OIP, OOC, STA)}
}
\value{
a ggplot2 object
}
\description{
Plot Biggest decrease in Refugee Population
}
\examples{
plot_reg_decrease(
  year = 2021,
  lag = 5,
  topn = 5,
  region = "The Americas",
  pop_type = c("REF", "ASY", "OIP")
)
}
