% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_evolution.R
\name{plot_reg_evolution}
\alias{plot_reg_evolution}
\title{Evolution over time}
\usage{
plot_reg_evolution(
  year = 2022,
  lag = 5,
  region,
  pop_type = c("REF", "IDP", "ASY", "OOC", "STA", "OIP")
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{lag}{Number of year to used as comparison base}

\item{region}{Character value with the related UNHCR bureau - when left null, it will display the whole world}

\item{pop_type}{Character value. Possible population type (e.g.: REF, IDP, ASY,   OIP, OOC, STA)}
}
\value{
a ggplot2 object
}
\description{
Display evoluation over time for specific population group (one or many) and 
defined number of years (lag)
}
\examples{
plot_reg_evolution(year = 2021,
                   lag = 5,
                   region = "Asia",
                   pop_type =  c( "REF", "IDP", "ASY", "OOC", "STA", "OIP"))
}
