% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_map.R
\name{plot_reg_map}
\alias{plot_reg_map}
\title{Plot a regional map}
\usage{
plot_reg_map(
  year = 2022,
  region = "Americas",
  topn = 5,
  pop_type = c("REF", "ASY", "OIP"),
  projection = "Mercator",
  maxSymbolsize = 0.25
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{region}{Character value with the related UNHCR bureau - when left
null, it will display the whole world}

\item{topn}{how many top countries to show..}

\item{pop_type}{Vector of character values. Possible population type 
(e.g.: REF, IDP, ASY, OIP, OOC, STA)}

\item{projection}{use a projection system - default is "Mercator"
 for instance this can be Bertin 1953 projection - 
https://visionscarto.net/bertin-projection-1953)}

\item{maxSymbolsize}{size in point to adjust for the maximum value
to display on the map}
}
\value{
a base plot
}
\description{
Plot a regional map
}
\examples{
plot_reg_map(  year = 2022,
                            region = "Asia",
                            topn = 5,
                            pop_type =  c("REF", "ASY", "OIP"),
                            projection = "Mercator",
                            maxSymbolsize = .25)


# plot_reg_map(  year = 2022,
#                             region = "WestAfrica",
#                             topn = 5,
#                             pop_type =  c("REF", "ASY", "OIP"),
#                             projection = "Mercator",
#                             maxSymbolsize = .25)

}
