% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_origin_dest.R
\name{plot_reg_origin_dest}
\alias{plot_reg_origin_dest}
\title{Plot Population Origin-Destination within the region}
\usage{
plot_reg_origin_dest(year = 2024, region = "The Americas")
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{region}{Character value with the related UNHCR bureau - when left null, it will display the whole world}
}
\value{
plot a ggplot2 object
}
\description{
Chord diagram showing Origin destination see - https://jokergoo.github.io/circlize_book/book/
}
\examples{
plot_reg_origin_dest(year = 2024, region = "Asia")
}
