% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_population_type_abs.R
\name{plot_reg_population_type_abs}
\alias{plot_reg_population_type_abs}
\title{Main country of origin - Absolute value}
\usage{
plot_reg_population_type_abs(
  year = 2022,
  region = "Americas",
  top_n_countries = 9,
  pop_type = "REF",
  show_diff_label = TRUE
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{region}{Character value with the related UNHCR bureau - when left null, it will display the whole world}

\item{top_n_countries}{Numeric value of number of main countries that the graph should display}

\item{pop_type}{Character value. Possible population type (e.g.: REF, IDP, ASY,   OIP, OOC, STA)}
}
\value{
a ggplot2 object
}
\description{
Main country of origin - Absolute value
}
\examples{
plot_reg_population_type_abs(year = 2022,
                              region = "Americas",
                              top_n_countries = 5,
                              pop_type = "REF",
                              show_diff_label = TRUE
                              )

plot_reg_population_type_abs(year = 2022,
                              region = "Americas",
                              top_n_countries = 5,
                              pop_type = "ASY",
                              show_diff_label = FALSE
                              )

}
