% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_prop_origin.R
\name{plot_reg_prop_origin}
\alias{plot_reg_prop_origin}
\title{Proportion of the population who are refugees, by country of origin}
\usage{
plot_reg_prop_origin(year = 2022, region = "Americas")
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{region}{Character value with the related UNHCR bureau - when left null, it will display the whole world}
}
\value{
a ggplot2 object
}
\description{
The proportion of a country???s population who become refugees is SDG indicator 10.7.4 
it consistutes is a useful way to identify the countries of origin producing the most refugees relative to their number of inhabitants.
}
\examples{
plot_reg_prop_origin(year = 2022,  region = "Americas")
}
