% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_rsd.R
\name{plot_reg_rsd}
\alias{plot_reg_rsd}
\title{Plot Chart on Refugee Status Determination}
\usage{
plot_reg_rsd(year = 2024, region, top_n_countries = 10, measure = "Recognized")
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{region}{Character value with the related UNHCR bureau - when left null, it will display the whole world}

\item{top_n_countries}{Numeric value of number of main countries that the graph should display}

\item{measure}{this can be either:
* Recognized
* ComplementaryProtection
* TotalDecided
* RefugeeRecognitionRate
* TotalRecognitionRate}
}
\value{
a ggplot2 object
}
\description{
Show the main host and origin countries based on number of decisions
}
\examples{
plot_reg_rsd(
  year = 2024,
  region = "The Americas",
  top_n_countries = 10,
  measure = "Recognized"
)


plot_reg_rsd(
  year = 2024,
  region = "The Americas",
  top_n_countries = 5,
  measure = "ComplementaryProtection"
)


plot_reg_rsd(
  year = 2024,
  region = "The Americas",
  top_n_countries = 10,
  measure = "TotalDecided"
)


plot_reg_rsd(
  year = 2024,
  region = "The Americas",
  top_n_countries = 10,
  measure = "RefugeeRecognitionRate"
)


plot_reg_rsd(
  year = 2024,
  region = "The Americas",
  top_n_countries = 10,
  measure = "TotalRecognitionRate"
)


# plot_reg_rsd(year = 2024,
#              region = "Europe",
#                         top_n_countries = 10,
#                         measure = "Recognized")
}
