% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solutions_long}
\alias{solutions_long}
\title{Persons availing a durable solution - long format}
\format{
A data frame with 19298 rows and 7 variables:
\describe{
  \item{\code{Year}}{character Year}
  \item{\code{CountryOriginCode}}{character Country of Origin Code isoA3}
  \item{\code{CountryAsylumCode}}{character Country of Asylum Code isoA3}
  \item{\code{CountryOriginName}}{character Country of Origin Name}
  \item{\code{CountryAsylumName}}{character Country of Asylum Name}
  \item{\code{Solution.type}}{character Type of Solution }
  \item{\code{Solution.type.label}}{character Type of Solution }
  \item{\code{Value}}{integer Number of person} 
}
}
\source{
\url{https://data.humdata.org/dataset/unhcr-population-data-for-world}
}
\usage{
solutions_long
}
\description{
Data collated by UNHCR, containing information about the number of forcibly displaced residing in World availing different types of solutions for refugees and IDPs such as repatriation or resettlement. Data is available since 1983.
}
\keyword{datasets}
