% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_ctryprez.R
\name{template_CtryPrez}
\alias{template_CtryPrez}
\title{Generate all country factsheet}
\usage{
template_CtryPrez(year = 2022, country_asylum_iso3c, folder = "Report")
}
\arguments{
\item{year}{Numeric value of the year (for instance 2022)}

\item{country_asylum_iso3c}{Character value with the ISO-3 character code of the Country}

\item{folder}{folder within your project where to put the generated report. 
Folder will be created if it does not exist}
}
\value{
nothing the file for the report is generated
}
\description{
Generate all country factsheet
}
\examples{
## generate for one country
# unhcrdatapackage::template_CtryPrez(year = 2022, country_asylum_iso3c = "CHL",   folder = "Report")

## Generate for a specific region
# region <- "Americas"
# 
# ## get all countries with more than 1000 Reported individuals
# ctr <- dplyr::left_join( x= unhcrdatapackage::end_year_population_totals_long, 
#                                 y= unhcrdatapackage::reference, 
#                                 by = c("CountryAsylumCode" = "iso_3")) \%>\%
#         filter(Year == year & 
#                 UNHCRBureau == region ) \%>\%
#         group_by( CountryAsylumName, CountryAsylumCode   ) \%>\%
#         summarise(Value = sum(Value) ) \%>\%
#         ungroup() \%>\%
#         filter( Value  > 1000 )
#   
#   for ( i in (1:nrow(ctr))) {
#     country_asylum_iso3c = ctr[i ,2 ]
#     template_CtryPrez(year = 2022, country_asylum_iso3c = country_asylum_iso3c,   folder = "Report") 
#   }
}
