% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_regfactsheet.R
\name{template_RegFactsheet}
\alias{template_RegFactsheet}
\title{Generate all country factsheet}
\usage{
template_RegFactsheet(year = 2022, region = "Americas")
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{UNHCRBureau}{Bureau that covers all the countrie factsheet to generate}
}
\description{
Generate all country factsheet
}
\examples{
# template_RegFactsheet(year = 2022, region = "Americas")
}
