% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dataset.R
\name{update_dataset}
\alias{update_dataset}
\title{Update the data on the package}
\usage{
update_dataset()
}
\value{
nothing just refresh the data .rda
}
\description{
Description
}
\examples{
update_dataset()
}
