% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{end_year_population_totals_long}
\alias{end_year_population_totals_long}
\title{Persons who are forcibly displaced, stateless and others of concern to UNHCR - reshaped}
\format{
A data frame with 103303 rows and 11 variables:
\describe{
  \item{\code{Year}}{character Year}
  \item{\code{CountryOriginCode}}{character Country of Origin Code isoA3}
  \item{\code{CountryAsylumCode}}{character Country of Asylum Code isoA3}
  \item{\code{CountryOriginName}}{character Country of Origin Name}
  \item{\code{CountryAsylumName}}{character Country of Asylum Name}
  \item{\code{Population.type}}{character Population.type: Refugees, Internally Displaced Persons,Asylum-seekers, Others of Concern to UNHCR, Stateless persons, Venezuelans Displaced Abroad }
  \item{\code{Value}}{integer Number of }
}
}
\source{
\url{https://data.humdata.org/dataset/unhcr-population-data-for-world}
}
\usage{
end_year_population_totals_long
}
\description{
Data collated by UNHCR, containing end-year stock population figures for forcibly displaced and stateless persons residing in World. Data is available since 1951.
}
\keyword{datasets}
