% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctr_disp_migrant.R
\name{plot_ctr_disp_migrant}
\alias{plot_ctr_disp_migrant}
\title{Plotting ratio Refugee/Asylum-seekers vs Migrant within a country}
\usage{
plot_ctr_disp_migrant(year = 2021, country_asylum_iso3c, top_n_countries = 10)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{country_asylum_iso3c}{Character value with the ISO-3 character code of the Country of Asylum}

\item{top_n_countries}{Numeric value of number of main countries that the graph should display}
}
\value{
a ggplot2 object
}
\description{
This chart provides insights on the relative weight of Forced Displacement for countries generating such type of displacement.
First, the size of the circle represents the number of Refugees, Asylum-seekers and other in need of International Protection
The vertical axis represent the Share of immigrants from the country in relation with all immigrants within the country. 
The higher is country in this axis, the more people from this country are the host country independently of the reason why they migrated.   
The horizontal axis displays the share of migration generated by forced displacement within each country of origin. 
The more a country is right, the more people from that country cam because of Forced Displacement.
}
\examples{
plot_ctr_disp_migrant(year = 2022,
                    country_asylum_iso3c = "MEX" ,
                    top_n_countries = 10)
# plot_ctr_disp_migrant(year = 2022,
#                     country_asylum_iso3c = "FRA" )
}
