% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctr_disp_migrant.R
\name{plot_ctr_disp_migrant}
\alias{plot_ctr_disp_migrant}
\title{Plotting ration Refugee/Asylum seekers vs Migrant within a country}
\usage{
plot_ctr_disp_migrant(
  year = 2021,
  country_asylum_iso3c = country_asylum_iso3c,
  pop_type = pop_type
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{country_asylum_iso3c}{Character value with the ISO-3 character code of the Country of Asylum}

\item{pop_type}{Vector of character values. Possible population type (e.g.: REF, IDP, ASY, OIP, OOC, STA)}
}
\description{
Plotting ration Refugee/Asylum seekers vs Migrant within a country
}
\examples{
plot_ctr_disp_migrant(year = 2021,
                    country_asylum_iso3c = "BRA",
                  pop_type = c("REF", "ASY", "OIP") )
}
