% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctr_population_type_per_year.R
\name{plot_ctr_population_type_per_year}
\alias{plot_ctr_population_type_per_year}
\title{Graph of population type per year}
\usage{
plot_ctr_population_type_per_year(
  start_year = 2015,
  country_asylum_iso3c = country_asylum_iso3c,
  pop_type = Population.type
)
}
\arguments{
\item{start_year}{Numeric value of first year}

\item{country_asylum_iso3c}{Character value with the ISO-3 character code of the Country of Asylum}

\item{pop_type}{Vector of character values. Possible population type (e.g.: REF, IDP, ASY, VDA, OOC, STA)}
}
\value{
Population type per year graph
}
\description{
Graph of population type per year
}
\examples{

plot_ctr_population_type_per_year(start_year = 2016,
                         country_asylum_iso3c = "PAN",
                          pop_type = c("REF", 
                                       "ASY", 
                                       "VDA", 
                                       "OOC",
                                       "STA",
                                       "IDP" )
                  )
}
