% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctr_pyramid.R
\name{plot_ctr_pyramid}
\alias{plot_ctr_pyramid}
\title{Population Pyramid}
\usage{
plot_ctr_pyramid(
  year = 2021,
  country_asylum_iso3c = country_asylum_iso3c,
  pop_type = pop_type
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{country_asylum_iso3c}{Character value with the ISO-3 character code of the Country of Asylum}

\item{pop_type}{Vector of character values. Possible population type (e.g.: REF, IDP, ASY, OIP, OOC, STA)}
}
\value{
a ggplot2 object
}
\description{
Population Pyramid
}
\examples{
# 
plot_ctr_pyramid(year = 2021,
                country_asylum_iso3c = "VEN",
                pop_type = c("REF", "ASY"))

}
