% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_map.R
\name{plot_reg_map}
\alias{plot_reg_map}
\title{Plot a regional map}
\usage{
plot_reg_map()
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{topn}{how many top countries to show..}

\item{region}{Character value with the related UNHCR bureau - when left null, it will display the whole world}

\item{pop_type}{Vector of character values. Possible population type (e.g.: REF, IDP, ASY, OIP, OOC, STA)}

\item{useBertin}{-Tells if the map should use Bertin 1953 projection - https://visionscarto.net/bertin-projection-1953)}
}
\value{
plot a ggplot2 object
}
\description{
Plot a regional map
}
\examples{
plot_reg_map()
}
