% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_population_type_per_year.R
\name{plot_reg_population_type_per_year}
\alias{plot_reg_population_type_per_year}
\title{Evolution over time}
\usage{
plot_reg_population_type_per_year(
  year = 2024,
  region = "The Americas",
  lag = 5,
  pop_type = c("REF", "ASY", "IDP", "OIP", "STA", "OOC")
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{region}{Character value with the related UNHCR bureau - when left null, it will display the whole world}

\item{lag}{Number of year to used as comparison base}

\item{pop_type}{Character value. Possible population type (e.g.: REF, IDP, ASY,   OIP, OOC, STA)}
}
\value{
a ggplot2 object
}
\description{
Display evoluation over time for specific population group (one or many) and 
defined number of years (lag)
}
\examples{
plot_reg_population_type_per_year(
  year = 2024,
  lag = 5,
  region = "The Americas",
  pop_type = c("REF", "ASY", "IDP", "OIP", "STA", "OOC")
)
}
