% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_treemap.R
\name{plot_reg_treemap}
\alias{plot_reg_treemap}
\title{Population group in the region}
\usage{
plot_reg_treemap(
  year = 2022,
  region = "Americas",
  pop_type = c("REF", "ASY", "IDP", "OIP", "STA", "OOC")
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{region}{Character value with the related UNHCR bureau - when left null, it will display the whole world}

\item{pop_type}{Vector of character values. Possible population type (e.g.: REF, IDP, ASY, OIP, OOC, STA)}
}
\value{
plot a ggplot2 object
}
\description{
Population group in the region
}
\examples{
plot_reg_treemap(year = 2022,  region = "Americas")
}
