% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solutions_residing}
\alias{solutions_residing}
\title{solutions_residing}
\format{
A data frame with 17127 rows and 9 variables:
\describe{
  \item{\code{Year}}{character Year}
  \item{\code{CountryOriginCode}}{character Country of Origin Code isoA3}
  \item{\code{CountryAsylumCode}}{character Country of Asylum Code isoA3}
  \item{\code{CountryOriginName}}{character Country of Origin Name}
  \item{\code{CountryAsylumName}}{character Country of Asylum Name}
  \item{\code{ResettlementArrivals}}{integer Number of Resettlement arrivals}
  \item{\code{Naturalisation}}{integer Number of Naturalisation}
  \item{\code{RefugeeReturns}}{integer Number of Refugee returns}
  \item{\code{IDPReturns}}{integer Number of IDP returns} 
}
}
\source{
\url{https://data.humdata.org/dataset/unhcr-population-data-for-world}
}
\usage{
solutions_residing
}
\description{
Data collated by UNHCR, containing information about the number of forcibly displaced residing in World availing different types of solutions for refugees and IDPs such as repatriation or resettlement. Data is available since 1983.
}
\keyword{datasets}
