% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_ctryfactsheet.R
\name{template_CtryFactsheet}
\alias{template_CtryFactsheet}
\title{Generate all country factsheet}
\usage{
template_CtryFactsheet(year = 2021, UNHCRBureau)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{UNHCRBureau}{Bureau that covers all the countrie factsheet to generate}
}
\description{
Generate all country factsheet
}
