% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_regprez.R
\name{template_RegPrez}
\alias{template_RegPrez}
\title{Generate all country factsheet}
\usage{
template_RegPrez(year = 2022, region = "Americas")
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{UNHCRBureau}{Bureau that covers all the countrie factsheet to generate}
}
\description{
Generate all country factsheet
}
\examples{
# template_RegPrez(year = 2022, region = "Americas")
}
