% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dataset.R
\name{update_dataset}
\alias{update_dataset}
\title{Update the data on the package}
\usage{
update_dataset()
}
\value{
nothing just refresh the data .rda
}
\description{
The function pull the data from https://data.humdata.org/dataset/unhcr-population-data-for-world and do a few adjustement before inserting them in the package
}
\examples{
#update_dataset()
}
